LayoutTests/imported/w3c/web-platform-tests/storage-access-api/storage-access-beyond-cookies.BroadcastChannel.sub.https.window-expected.txt


FAIL Verify StorageAccessAPIBeyondCookies for Broadcast Channel assert_equals: Storage Access API should be accessible and return first-party data expected "HasAccess for BroadcastChannel" but got "Unable to load handle in cross-site context for all"
